Revert "Fixed StatusChanged messages with invalid NetConnectionStatus"
This reverts commit 72f4b1cf1c7b9189b7356793ab01782ac8e2531b.

+++ Lidgren.Network/Connection/NetConnection.cs
@@ -131,7 +131,7 @@ namespace Lidgren.Network
                 info.EnsureCapacity(4 + reason.Length + (reason.Length > 126 ? 2 : 1));
                 info.SenderConnection = this;
                 info.SenderEndPoint = RemoteEndPoint;
-                info.Write((byte)Status);
+                info.Write(Status);
                 info.Write(reason);
                 Peer.ReleaseMessage(info);
             }
